3/requirements.txt
Flask
Authlib
Flask-Login
python-dotenv
